+++ packages/lit-html/src/hydrate.ts
@@ -13,14 +13,13 @@
  */
 
 // Type-only imports
-import {TemplateResult, DirectiveResult} from './lit-html.js';
+import {TemplateResult} from './lit-html.js';
 
 import {
   noChange,
   EventPart,
   NodePart,
   PropertyPart,
-  NodePartInfo,
   RenderOptions,
   ATTRIBUTE_PART,
   $private,
@@ -30,6 +29,7 @@ const {
   _TemplateInstance: TemplateInstance,
   _isIterable: isIterable,
   _isPrimitive: isPrimitive,
+  _resolveDirective: resolveDirective,
 } = $private;
 
 type TemplateInstance = InstanceType<typeof TemplateInstance>;
@@ -213,7 +213,8 @@ const openNodePart = (
         value = result.value;
       }
       (state.part._value as Array<NodePart>).push(part);
-    } else { // state.type === 'leaf'
+    } else {
+      // state.type === 'leaf'
       // TODO(kschaaf): This is unexpected, and likely a result of a primitive
       // been rendered on the client when a TemplateResult was rendered on the
       // server; this part will be hydrated but not used. We can detect it, but
@@ -239,12 +240,7 @@ const openNodePart = (
   // 6. Iterable
   // 7. nothing (handled in fallback)
   // 8. Fallback for everything else
-  const directive =
-    value != null ? (value as DirectiveResult)._$litDirective$ : undefined;
-  if (directive !== undefined) {
-    part._directive = new directive(part as NodePartInfo);
-    value = part._directive!.update(part, (value as DirectiveResult).values);
-  }
+  value = resolveDirective(part, value);
   if (value === noChange) {
     stack.push({part, type: 'leaf'});
   } else if (isPrimitive(value)) {
@@ -282,7 +278,9 @@ const openNodePart = (
     } else {
       // TODO: if this isn't the server-rendered template, do we
       // need to stop hydrating this subtree? Clear it? Add tests.
-      throw new Error('Hydration value mismatch: Unexpected TemplateResult rendered to part');
+      throw new Error(
+        'Hydration value mismatch: Unexpected TemplateResult rendered to part'
+      );
     }
   } else if (isIterable(value)) {
     // currentNodePart.value will contain an array of NodeParts
@@ -383,7 +381,12 @@ const createAttributeParts = (
         instancePart instanceof EventPart ||
         instancePart instanceof PropertyPart
       );
-      instancePart._setValue(value, state.instancePartIndex, noCommit);
+      instancePart._setValue(
+        value,
+        instancePart,
+        state.instancePartIndex,
+        noCommit
+      );
       state.templatePartIndex++;
       state.instancePartIndex += templatePart._strings.length - 1;
       instance._parts.push(instancePart);