+++ packages/lit-html/CHANGELOG.md
@@ -62,6 +62,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added `renderBefore` to render options. If specified, content is rendered before the node given via render options, e.g. `{renderBefore: node}`.
 - Added development mode, which can be enabled by setting the `development` Node exports condition. See `README.md` for more details.
 
+* Added `unsafeStatic()`, which allows template authors to add strings to the
+  static structure of the template, before it's parsed as HTML.
+
 ### Fixed
 
 - All usage of `instanceof` has been removed, making rendering more likely to
@@ -73,7 +76,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [Breaking] The `templateFactory` option of `RenderOptions` has been removed.
 - [Breaking] TemplateProcessor has been removed.
 - [Breaking] Symbols are not converted to a string before mutating DOM, so passing a Symbol to an attribute or text binding will result in an exception.
-- [Breaking] The `until`, `asyncAppend` and `asyncReplace` directives are not implemented.
+- [Breaking] The `asyncAppend` and `asyncReplace` directives are not implemented.
 
 ## [1.3.0] - 2020-08-19